Added ability to specify explicit pidfile directory
The redis_install resource and the default recipe can now have their pidfile
directory explicitly set via a piddir attribute.  This attribute is set to
nil by default, allowing the current base_piddir-derived functionality to
continue being used as the default, but giving the user the ability to
opt in to using an explicit pidfile directory.